moving tests to PageObjects folder for now - related to #107
